pyproject.toml
[project]
name = "evaluation-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "google-generativeai>=0.8.5",
    "ipython>=9.4.0",
    "langchain>=0.3.27",
    "langchain-community>=0.3.27",
    "langchain-core>=0.3.74",
    "langchain-groq>=0.3.7",
    "langchain-ollama>=0.3.6",
    "langgraph>=0.6.5",
    "langgraph-cli[inmem]>=0.3.6",
    "uvicorn>=0.35.0",
    "watchdog>=6.0.0",
]
